Create book_recomendations.py
this is a book recomendations program i created for a job interview assignment. it basicaly recommends user books based on the book that he likes. it uses a super small database i created in a hury so if you would like to use it with a different one you might have to change the code a bit. however everything is explained in comments and doctsring.